requirements.txt
fastapi[standard]
pydantic
requests
openai
python-dotenv
sqlite3